create MathUtil to compartmentalize statistics
